cleaned code structure and stubbed player movement
Board now has the player position, since in level it couldn't be
accessed either by the things that needed to read it or the things that
update it. ALso has a stubbed function to move the player, invoked when
a revealed cell is clicked. Some other code reorganization as well.